[Move] FavorCompositionalLayout 관련 파일 패키지로 이동 및 접근제어자 수정
